refactor: separate the controller from roles
