+++ API/CapturedStyles/src/main/java/co/edu/ue/controller/admin/ProductController.java
@@ -1,0 +1,4 @@
+package co.edu.ue.controller.admin;
+
+public class ProductController {
+}
